Testing the beep infrastructure in a loud datacenter, this is my conclusion:
 - Beeping through ctrl-g using the terminal cannot be heard
 - Normal beeping (at 750Hz) cannot be heard
 - Beeping between 2000Hz and 2200Hz is best, although you
   have to listen very well.
 - Some systems have a very low volume speaker and it's hard
   to capture the beep if you don't know what to expect.

So I guess we should not (only) rely on beeping, even though
it works fine in other environments. For our intended use-case
there's no loud datacenter in most cases.

We may want to use the frontend ID led during 'rear udev' run,
as it is very visible on the system (usually very close to the
USB port).

Or we have to look into some way to cut the power of the USB
device in order to turn off the light of the USB device.
(Windows can do it, haven't seen this on Linux yet)

Another crazy idea is to eject/retract the CD drive ;-) Or
in older servers, read from the (empty) floppy drive ;-)