layout: show partitions of multipath devices in the layout file.
Partitions on device mapper devices in general do not show up as children of the device in sysfs. Instead, they get their own entry under /sys/block (and only after a partprobe/reboot).

That entry lacks a start attribute, so we fall back to the code used during migrations when recreating the partitions.