Improve the BUG BUG BUG! error string
I was considering getting rid of the "BUG BUG BUG!" string, however
some projects might be testing for this string, so we may have to
keep it around (for now).

However, the message to report the issue was improved, including
a link to the issue-tracker, a reference to the log file and
a HINT to reproduce the problem using the -d option.