Ensure we have the new partitioning on the device
I opted to add the '-s' option to partprobe so that we can verify during debugging whether the new partitioning is how we expect it to be. Especially when this change does not fix all cases, it could be useful to know what is wrong at this stage.

This fixes #147