Report to the user when ROOTFS_DIR's filesystem is mounted 'noexec'
This change adds a function to find the filesystem name of a path, and also adds a test to see if the ROOTFS_DIR's filesystem is mounted noexec. This is required in order to report that we have skipped the chroot bash test.

This change does not implement a configurable BUILD_DIR or TMP_DIR !

This closes issue #150