Revert some of the SPEC file changes, and make the crontabs dependency a Red Hat thing.
There is no reason to define a non-existing Source1 that we create ourselves. The cron-job is created in %prep so it's not a real "source" in fact.

There is also no reason to change "rear" by the package name, I know some people think that is a best practice, but it makes it more error-prone and doesn't help in readability. And you can't easily grep for filenames because of it. Explicit trumps implicit. Less abstraction wins too.