Remove restriction of supporting only 9 partitions from find_partition_number()
So the assumption is that partition device /dev/mapper/36001438005deb05d0000e00005c400001 is not a real example. And if it was, we don't support that anymore. I guess time will tell...

See discussion #263 for more information.

PS StopIfError does not stop Relax-and-Recover immediately, it may continue and cause another StopIfError before it really bails out. This is confusing because the second error-message may be caused by the first error, but it could be the only error that the user is looking at...