Fixed test if filesystem is mounted with user_xattr. Was not working with btrfs.
user_xattr is the default in btrfs and btrfs does not even allow the mount
option user_xattr (btrfs: unrecognized mount option 'user_xattr') the
previous test did report that the btrfs filesystem is not
mounted with user_xattr.

This new test tries to actually create a file with extended attributes.

Tested it with btrfs and tmpfs.