added usr/share/rear/layout/prepare/GNU/Linux/SUSE_LINUX/14_include_mount_filesystem_code.sh script to override the mount_fs function (therefore we start with number 14 instead of 13 so we are sure it gets read after the default mount_fs() function defined in usr/share/rear/layout/prepare/GNU/Linux/13_include_mount_filesystem_code.sh)
This code is coming from J Meixner (see issue #497) read116 fork from SuSe. We are just moving the code from Johannes back into the main stream of rear, but now we extracted the btrfs mount code away from 13_include_filesystem_code.sh into a specific script for SUSE_LINUX called 14_include_mount_filesystem_code.sh;

The purpose is to get one script in the future that could handle all types of btrfs (which I doubt). We also need to think on a Debian specific script to (but that is for later - see issue #408)

Issues #233, #497 and #408