Change TSM default filespaces restored
See Issue #528

The TSM_RESTORE_FILESPACE_NUMS defaulted to all known filespaces
which might not be desired if the filesystem was not recovered
as part of the rescue image. This change still allows to choose
any filespaces that TSM knows about but will default to only those
that are part of the rescue image.

- Created new TSM_FILESPACE_INCLUDED_NUMS array which is now used
  as the default for the TSM_RESTORE_FILESPACE_NUMS
- Populated the MOUNTPOINTS_TO_RESTORE variable with those
  fileystems that were excluded from the rescue image because TSM
  does not recreate those by default