Create 36_teaming.sh
This script will add the teaming function to rear. Because teaming runs in user space we need more than just the kernel and the drivers to get this kind of interfaces online. We would need also daemons and tools. Because of this and for simplify this function, routing and ip addresses will be attached to the first available teaming-member (output by teamdctl).