Grub2: Using search --fs-uuid instead of hardcoded root='hd0,msdos1'
Actually we found that ppc64 and ppc64le are not aware of
"hd0,msdos1", current grub2 rescue does not work without the above option in ppc64 and ppc64le. Since this search option is used even in x86 grub2 distro menu, so it would be better to handle as platform independent option.