Let /bin/ldd detect *.so with relative paths
This allows the ldd test in 980_verify_rootfs.sh to work properly
for binaries which have relative paths to their *.so files.
See https://github.com/rear/rear/pull/1560 for discussion regarding
/opt/fdrupstream/uscmd1.

This commit includes the use of 'bash --login -c' in order to allow
the 'cd' command to work properly in the chroot call.