Issue #1913 - NETFS+tar backup always succeeds even if there was an error
The ReaR code is making the assumption that the $BACKUP_PROG return
value is returned from the child process. This is true for **rsync** and
unsupported methods, but not for **tar** because **tar** uses the
following pipes:

  tar ... | <encrypt> ... | <splitter>

which ends up for basic case with the following pipes:

  tar ... | cat | dd of=backup.tar.gz

In fact, the return code is hence the one for **dd**.

When there is an error, e.g. ENOSPC, **dd** returns 1, causing ReaR to
believe **tar** returned 1, which is non-critical (1 means there was
file modifications while reading them).
Also, if **tar** suffered an issue, e.g. an invalid parameter was passed
to **tar**, then **dd** would return 0, causing a fake success to be
returned.

This code fixes all these issues by returning the exit code of the last
pipe failing (can be **dd**, or **cat**, or **tar**).
It also enhances the exit code processing code, by displaying the
command's output if possible.

Signed-off-by: Renaud Métrich <rmetrich@redhat.com>