Fix for 'error:unrecognized number' when booting ISO on PPC
Issue is caused by incorrect, according to PAPR specification,
bootinfo entity parsing code in grub2.

In PAPR spec "B.4.1.6.1 Bootinfo Entities" partition and path
separator (,) is optional and if not present bootinfo may be
treated as directory and filename components unless it begins
with number:

  &device;:[&partition;][,]&directory;&filename;

If (,) is missing and directory begins with a hexadecimal number
GRUB2 treats this as partition number immediately followed by a
directory/filename.

Note that with (,) present in bootpath issue isn't in the case
as GRUB2 returns empty string for GRUB_PARSE_PARTITIONS. This
effectively meaning that GRUB2 code requires (,) as mandatory.

This is a GRUB2 bug, but until it is fixed, we can make the message disappear by specifying a comma in bootinfo.txt.

Signed-off-by: Renaud Métrich <rmetrich@redhat.com>