Update 470_safe_filesystems.sh
It is of additional use to retrieve the filesystems in NSR_CLIENT_MODE; it does not harm since it is "read-only" on EMC Networker server. Due to saving the save sets filesystem information in $VAR_DIR/recovery/nsr_paths within the recovery image it is also possible to retrieve/read them during a recovery process.