NBU: exclude '/usr/openv/netbackup/logs' directory instead of files in directory only
When using the NetBackup method, the files in the logs directory are
excluded because they are useless. However, if, for some reason
/usr/openv/netbackup/logs is a symlink to some other directory (a user
customization), then ReaR will fail to create the ISO, due to having a
skeleton directory for NBU in ReaR, causing the 'tar | untar' of
COPY_AS_IS items to fail.

Signed-off-by: Renaud Métrich <rmetrich@redhat.com>