Fix including of multipath disks in backup
The AUTOEXCLUDE_DISKS logic (320_autoexclude.sh) traverses filesystems
and tries to determine the associated underlying disks that should be
kept. When having a filesystem fs:/A that is present on a multipath
device MP and the root filesystem fs:/ that is present on a normal disk
D, the code first established that fs:/A depends on fs:/ resulting in
marking D as used and then short-circuited processing of multipath
devices. This resulted in MP not being considered as a used device and
removed from the backup.

The patch fixes this problem by removing the short-circuit logic in
find_disk_and_multipath() so multipath devices get processed too (unless
AUTOEXCLUDE_MULTIPATH is true).

Fixes #2236.