Split GRUB2_MODULES into GRUB2_MODULES and GRUB2_MODULES_LOAD
The original semantics of GRUB2_MODULES (what modules to include in the Grub2
core image) is now carried by GRUB2_MODULES_LOAD. GRUB2_MODULES now determines
what modules to install into the embedded ramdisk, where they are not
prelaoaded, but available for manual loading (insmod) or autoloading.

The change is done for symmetry with MODULES and MODULES_LOAD, which have the
same semantics for kernel modules.

The previous semantics (GRUB2_MODULES determined both what to install and what
to load) was yielding unexpected results: when using this variable to extend
the set of modules to load, one was simultaneously reducing the set of modules
to install. As a result, commands like "ls" (which usually is not part of the
core image, but is autoloaded) were not available.