Update stale.yml
Longer periods of days to be more on the safe side that
there are at laset two full working weeks and three full month
because it could look unfriendly to the user when issues get
automatically closed a bit sooner than two weeks or three month.
Added new label "severe improvement" that is excluded from
automatic close to keep issues open that are about heavyopen for an 
improvements that are required for long term success of ReaR
i.e. in particular improvements that need years to be implemented
where longer periods of inactivity are normal and expected.