Update 100_copy_as_is.sh
Filter out duplicate entries in COPY_AS_IS
but keep the ordering of the elements
i.e. only the first occurrence of an element is kept
and remove duplicates in the copy_as_is_filelist_file
with 'sort -u' because here the ordering does not matter
cf. https://github.com/rear/rear/issues/2377