Error out if files greater or equal 2GiB in ISO
Error out when files greater or equal ISO_FILE_SIZE_LIMIT should be included in the ISO.
There is a limit of the ISO 9660 file system that is 2GiB or 4GiB according to
https://en.wikipedia.org/wiki/ISO_9660#The_2/4_GiB_file_size_limit
To be on the safe side we use by default a 2GiB limit.
2GiB is at least the actual limit when ebiso is used, cf.
https://github.com/gozora/ebiso/issues/12
Under normal circumstances files greater or equal 2GiB should not appear in the ISO.
An exception is when the backup is included in the ISO with BACKUP_URL=iso://
where the backup archive file (e.g. backup.tar.gz) could be greater or equal 2GiB.
The user might adapt ISO_FILE_SIZE_LIMIT provided he verified that "rear recover"
actually works in his particular environment even when there are files in his ISO
(in particular backup.tar.gz) that are actually greater than the default 2GiB limit.
When there is a 2GiB file size limit a backup.tar.gz that is greater than 2GiB
will get corrupted in the ISO so backup restore via "rear recover" would fail.
Also the ReaR recovery system initrd could become greater or equal 2GiB
(e.g. because of accidentally too much in COPY_AS_IS) which is very likely an error
so we error out when files greater or equal ISO_FILE_SIZE_LIMIT should be included in the ISO.