New ONLY_INCLUDE_DISKS variable
This array variable enables to only process specified disks when
building the disklayout.conf file.  This is particularly useful when
having a lot of multipath devices which are slow to scan but anyway
excluded in the end.

This is for experts only, since it may end up creating an inconsistent
disklayout.conf file. In particular the EXCLUDE_MOUNTPOINTS variable is
not populated automatically because doing so would likely break things.

Usage example:

ONLY_INCLUDE_DISKS=( sda sdb )
or
ONLY_INCLUDE_DISKS=( /sys/block/sda )

Signed-off-by: Renaud Métrich <rmetrich@redhat.com>