Teach get_child_components() to skip device types
Sometimes not all ancestors of a given filesystem are desired.

If we want to install a bootloader on a disk containing /boot,
we need to locate the partition that /boot resides on. We do that using
layout dependencies.
Unfortunately, dependencies also include dependencies between
filesystems: a filesystem depends on filesystems that contain its
mountpoint. We are not interested in this kind of dependencies,
otherwise we end up installing the bootloader on the disk containing the
root filesystem (if different from the /boot disk).

In order to solve this, add an optional argument to
the get_child_components() and the find_*() functions in
layout-functions.sh. It is a list of dependency types that should be
ignored when traversing the dependency tree. Unlike the second argument
to get_child_components(), it affects not only the final results but
also the intermediate tree nodes.

In order to solve the problem above, this argument can be set to "fs"
and dependencies between filesystems will be ignored.