Update 400_save_directories.sh
Fixed detection when a directory is already listed in the directories_permissions_owner_group file:
Before it was: grep "^$directory" "$directories_permissions_owner_group_file" 1>&2
Now it is: grep -q "^$directory " "$directories_permissions_owner_group_file"
The crucial change is the added trailing space in "^$directory " that avoids
e.g. '/foo' is found when only '/foobar ...' is already listed in that file.
Fortunarely the directories in FHSdirectories are properly sorted
so that things worked well by luck.