Remove code that depended on the current stage
Since a8fdc445d0d6f7f9184dc6633817928f96aae9bc the handling of
BACKUP_URL and OUTPUT_URL is separate and the code is not shared between
backup and output stages. Also, the code that handles the lockfile is
different anyway. This allows some simplification by removing the
indirection based on the current stage that originally allowed sharing
the code.

Making this code generic and shared between the stages might be a good
idea, but would have to be done a bit differently.