Move mountpoint removal/creation into u/mount_url
All callers of mount_url() create a temporary mount point for it, so we
can move the mount point creation there, together with all the exit task
logic, and the corresponding removal into umount_url().

Functional change: mount point does not get created and removed for URLs
that do not need to be mounted. This mount point would be unused and if
not, it would be a bug (nobody would save files that appear there, so
they would get lost).

Bug fix: call the filesystem-specific umount handling for davfs and
sshfs (it used to be ignored).