Detect unmountable URL with scheme_supports_filesystem
If scheme_supports_filesystem returns false, the URL is definitely not
mountable. Use this instead of keepeing a long list of unmountable URL
schemes in mount_url() and umount_url().

There are some URLs that allow filesystem accces even if we don't
mount and umount them (file: and sometimes iso: ). They need to be
handled as special case. Still, this reduces the amount of hardcoded
schemes quite a lot.