Suggest a safe way to remove build area
When the build area is left for debugging purposes, we tell the user
that they should remove it later. Since the output filesystem with
valuable old backups or backups from other machines may be still mounted
at $BUILD_DIR/outputfs, tell the user to use rm -Rf --one-file-system
instead of just rm -Rf. This prevents data loss if the user follows the
intruction literally.

All supported distributions now support the --one-file-system option, so
there should be no problem with this.