If checking integrity abort restore on rsync error
For BACKUP=NETFS, when BACKUP_INTEGRITY_CHECK is set, the restore
process errors out when the backup program (tar, rsync) returns an error
(nonzero exit code).

Do the same for BACKUP=RSYNC. This allows the user to catch possibly
broken restored systems early. Related issue: #2612

Not done by default: the reasoning is that it is better to continue the
recover process in order to get the system into a basically usable
state (the problem may be with restoring some unimportant file).

This does not mean that BACKUP_INTEGRITY_CHECK is now fully supported
for BACKUP=RSYNC: checking of md5sums is not there.