Check fake super support only when RSYNC_PROTO=ssh
If RSYNC_PROTO=rsync, there is no way to know what the server supports,
the "$TMP_DIR/rsync_protocol" file gets created only when using ssh.
Performing this check for rsync protocol is useless anyway:
--rsync-path="rsync --fake-super" will not do the right thing, because
--rsync-path is not used when talking to the rsync daemon. (I would
assume that -M--fake-super would work, but in my tests it does not work
either, unfortunately.)
Fixing the fake super check for the RSYNC_PROTO=rsync case was probably
the intent of 3548430302cef3a8445de970eb763da33d9f6e8e, but it can't
work this way and fixing it properly will be more work (if it can be
done at all).