Set TMPDIR to /var/tmp if unset and export it
The default path ( /tmp ) is not suited for ReaR, because ReaR needs
lots of space. file-hierarchy(7) recommends to use /var/tmp as default
for such programs.

If the user sets TMPDIR explicitly, it still takes precedence.

The variable is exported because it influences the behavior of utilities
that ReaR invokes, in particular mktemp(1).

See the discussion in issue #2654 .