Remove code that unset TMPDIR in rescue.conf
This code probably has not worked since commit 0022063 (PR #2633),
because rescue.conf is sourced only when the build area has been
already created by invoking mktemp. It is also unneeded, because since
that commit, setting TMPDIR in local.conf (which was the initial concern
that motivated this code) does not work anyway, as discussed in #2654.