Require bootlist & ofpathname if needed
As reported by @rmetrich, the ofpathname binary is called by
grub2-install. Therefore, it is required in the rescue system in order
to make the recovered system bootable, except for the PowerNV (Not
Virtualized - bare metal) case. Under PowerVM, we also need the bootlist
executable to make the system bootable.

Add those two binaries to REQUIRED_PROGS instead of PROGS under the
appropriate conditions.

Do not handle PowerNV for ppc64: according to
89ddb9fc17adb022ce9c10be0c3e5b835ba139d7, only ppc64le Linux can run in
PowerNV.

(Wouldn't it be simpler though to symlink Linux-ppc64.conf to
Linux-ppc64le.conf? Their differences seem to be minimal.)