Move a check for block device to the layout script
After this change, nothing in 100_include_partition_code.sh requires the
actual hardware anymore - only the saved layout. This enables running
the script outside of the rescue environment for testing, if one has
the saved layout data, and thus facilitates testing and experimenting a
lot.

Functional change: if the disk is not found, it is now the layout script
that aborts with BugError, not whole ReaR.