Update 300_map_disks.sh
Better wording in user message - now it is
"Using /dev/sdX (the only available of the disks) for recreating /dev/sdY"
instead of what it was before
"Using /dev/sdX (the only appropriate) for recreating /dev/sdY"
because "appropriate" sounds as if /dev/sdX is the right one to recreate /dev/sdY
but acutally /dev/sdX could be inappropriate to recreate /dev/sdY (e.g. much too small)
but /dev/sdX is the only diks that is left available of all existing disks
so /dev/sdX is the only possible choice that is left to recreate /dev/sdY
under the conditions of the currently already existing mappings
and the exclusions from possible mapping targets.