Update 400_copy_modules.sh
Fix function modinfo_filename:
Output the original module_filename also when it is a symlink
because in the code below 'cp -L' copies the symlink taget content
as a new regular file with file name as the name of the symlink
so the copied content can be still found under its original name,
cf. https://github.com/rear/rear/issues/2677#issuecomment-997859219
Additionally in case of MODULES=( 'all_modules' ) also use 'cp -L' to
copy the actual content to avoid dangling symlinks in the recovery system.