Verify checksums after file restore from backup
Configuration files can get out-of-sync with the recreated layout.
This leads to inconsistent configuration, e.g. filesystem not mountable
at boot, if they are mounted by UUIDs, and failures to migrate UUIDs in
files (if needed).

To detect this problem, use the saved md5sums of configuration files
(originally used for "checklayout") to verify the restored files and
inform the user in the case of a mismatch.

No attempt is made to repair the problem; the user is informed, but they
must fix the problem manually.