Save hashes of all files that need ID migration
If a file may need a migration of UUIDs, then it is very likely prone to
get out-of-sync with the layout if it gets modified. Add all such files
to the list of files whose hashes are recorded in the saved layout to
detect such problems.

Introduce a variable called FILES_TO_PATCH_PATTERNS to hold the list of
these files (shell glob patterns, actually), as it is now needed in
multiple places.

Would it make sense to just remove [e]tc/mtab from
FILES_TO_PATCH_PATTERNS? The file used to be maintaned dynamically by
mount(8) even before it became a symlink. Why has it been ever useful to
patch the IDs or device names in it, if it gets regenerated on every
boot anyway?