Add /usr/libexec/vi to the rescue ramdisk
On recent Fedora and RHEL 9, vi is a shell script that calls vim or
/usr/libexec/vi. If neither is installed, the vi command is not useful.

One could add vim as well, but I am afraid that it would bloat the
rescue ramdisk in case vim is compiled with a rich set of features.

(Should there be a Fedora-specific configuration file? There is already
SUSE_LINUX.conf and Ubuntu.conf .)