Replace mkinitrd with dracut on Fedora and RHEL
For a long time, mkinitrd was a wrapper script around dracut in Fedora
and RHEL that maintained backwards compatibility with the real (but
rather obsolete) mkinitrd.

Dracut 054 removed the mkinitrd wrapper script [1] so now we have to
execute dracut directly.  Fortunately, ReaR did not use almost anything
from mkinitrd's interface (with the exception of --with) so the transition
is quite straightforward and should not introduce any regressions because
dracut is the default initrd generator since RHEL 6.

Without this patch the initrd regeneration in the finalize stage of recovery
fails on RHEL 9 and Fedora 33 or newer.

[1] https://github.com/dracutdevs/dracut/pull/1285