Fix parsing of rsync URLs
The original code did not cope well with rsync protocol and
multi-compontent paths (like module/path instead of just module). It
also did not cope well with trailing slashes at the end of paths when
using the rsync protocol and old-style ssh URLs with paths that do not
have a leading slash ([USER@]HOST:PATH instead of [USER@]HOST:/PATH).