Pass -y to lvcreate to stop asking for input
Revert "Use fail-safe 'yes' pipe for "lvm lvcreate""

This reverts commit 9d6ce289938092834425cb1c27658500f68bc69f.

The -y argument is much simpler and less error-prone than a 'yes' pipe
(no error termination because of SIGPIPE, for instance). But we lose
support for SLES11, where lvcreate does not know the -y option.

'yes' remains among REQUIRED_PROGS, it can be useful in the rescue
system. (It could be theoretically moved to PROGS, because it is not
required anymore. This could be done separately if really desired, but
for now we prefer to keep it in REQUIRED_PROGS, sooner or later
something will start requiring it.)

See https://github.com/rear/rear/pull/2834