Remove the source URL from generated RPM spec
Remove the sourceforge.net URL from the Makefile recipe that generates
the Source tag in the RPM spec in the `dist` target. We don't use
sourceforge.net anymore and the Source points to a locally-generated
tarball, so the URL has not been valid anyway.

If $(distversion) == $(version) (when OFFICIAL is set), we could keep
the Source tag unchanged - this is true of most of the sed
transformations in this recipe.