Validate disk when saving layout
Introduce function is_disk_valid, which performs checks for disk
usability beyond validating the device name. In some cases the device
name may be valid, but there are no data, typically because it is a
drive with removable media and there is no medium in the tray. Happens
typically with card (e.g. SD card) readers with empty slot.

This is a normal occurrence, so do not Error out, only display a message
and skip the device.