Revert "Eliminate a %define from RPM spec"
While defining a macro on the command line is more elegant, not having
the actual value of 'rpmrelease' in the spec file causes it to be lost
when one builds the resulting SRPM manually, because rpmbuild is not
invoked with the correct command line (unless one takes care to pass the
same parameters to rpmbuild that "make rpm" would use).

Having "make rpm" produce different RPMs (different NVR) than manually
building the SRPM generated by "make srpm" is confusing, so restore the
former behavior.

This reverts commit f8c77f6831de09802038f60acd2ba2a0ba6fb864.