Revert "Update 200_partition_layout.sh"
This reverts commit 0a1d634ed15500bb21f37ac1bbb11c8a4bb11545.

We now skip disks with no data (like when there is no medium), so
incomplete disk entries (without partition type) should not occur
anymore. Restore the code that aborted when such disks were encountered.
Incomplete entries should not be allowed to occur, as they could confuse
the layout restoration code. Moreover, the layout restoration wipes
all disks in the layout, so if during layout restoration there happens
to be a medium in the drive that was empty during layout save, the data
on the medium would get overwritten and lost. And if there is not medium,
the layout recreation script would fail.

See the discussion at https://github.com/rear/rear/issues/2958#issuecomment-1479588829