Make sure rescue contains all COPY_AS_IS files
The files copied as-is in the rescue are copied using a 'tar -c | tar
-x' command. It may happen that inodes are shrinking during the copy
(e.g. /dev/mqueue/nnsc (HPE device)), which can cause the 'tar -x'
command to stop extracting the next files, due to the padding zeros
inserted in the shrank inode.
The solution is to add '-i' option when extracting, to continue the
processing.