Add XFS-specific mount code to layout restoration
The mount command displays all mount options for a filesystem, including
those that are not explictitly set in fstab, and ReaR saves them to disk
layout. In the case of XFS, some of them can be harmful for mounting the
filesystem during layout restoration:

The logbsize=... mount option is a purely performance/memory usage
optimization option, which can lead to mount failures, because it must
be an integer multiple of the log stripe unit and the log stripe unit
can be different in the recreated filesystem from the original
filesystem (for example when using MKFS_XFS_OPTIONS, or in some exotic
situations involving an old filesystem, see GitHub issue #2777 ). If
logbsize is not an integer multiple of the log stripe unit, mount fails
with the warning "XFS (...): logbuf size must be greater than or equal
to log stripe size" in the kernel log (and a confusing error message
"mount: ...: wrong fs type, bad option, bad superblock on ..., missing
codepage or helper program, or other error." from the mount command),
causing the layout restoration in the recovery process to fail.

Wrong sunit/swidth can cause mount to fail as well, with this in the
kernel log: "kernel: XFS (...): alignment check failed: sunit/swidth vs.
agsize".

Therefore, remove the logbsize=... and sunit=.../swidth=... from XFS
mount options before mounting the file system.

(Another possibility would be to remove them already when saving the
layout, in layout/save/GNU/Linux/230_filesystem_layout.sh, but I decided
to follow the example of btrfs here.)