Save LVM pool metadata volume size in disk layout
Instead of letting LVM use the default pool metadata volume size when
restoring a layout with thin pools, use the size from the original
system. The pool metadata size will be saved in disklayout.conf as the
new key "poolmetadatasize" on the thin pool LV.

This makes the layout of the recovered system closer to the original
system.

Prevents some cases of recovery failures during layout restoration: if
the original system used a non-default (in particular, smaller) metadata
volume size, and the space in the VG was fully used, there would not be
enough space in the VG for recovery of all LVs (as layout restoration
would create a larger pool metadata size than before and if restoring to
disks of the same size, the space will be missing elsewhere).