Do not mount /sys if already mounted
Newer versions of systemd (starting with Fedora 39) seem to mount /sys
themselves. Mounting it again leads to errors on the recovery system
startup (startup scripts failing with status=219/CGROUP ), see
https://github.com/rear/rear/issues/3017.

Check if /sys is already mounted using the `mountpoint` tool and mount it
only if it is not.

Do the same for the other system mountpoints like /proc, /dev, /dev/pts.
Not sure if they suffer from the same problem, but they probably could.

N.B. the `mountpoint` command is already among REQUIRED_PROGS.