Take AUTOEXCLUDE_PATH into account in layout doc
Since PR #2261, /media is included in the default AUTOEXCLUDE_PATH. This
means that a filesystem mounted on a directory like /media/backup will
not be included in the layout. This makes the example invalid, as it
describes how a filesystem mounted under /media/backup will be included
in the layout and what manual steps are needed to exclude it.

To preserve the validity of the example, change all paths in the example
from /media/backup to /backup.

Take this opportunity to also describe the AUTOEXCLUDE_PATH variable
among other autoexclusions.