Don't copy EXCLUDE_BACKUP/RESTORE to layout doc
EXCLUDE_RESTORE and EXCLUDE_BACKUP are irrelevant for layout file and
disk layout recreation. They merely influence file backup or restore
(for supported backup methods). This guide is about the recreation of
the disk layout. Therefore, omit the variables from here. They are not
used in the doc anywhere anyway after showing a copy of their
description originating from default.conf. So, why to show it at all?