layout doc: move autoexcludes description
I believe that "This behavior is controlled by the
+AUTOEXCLUDE_DISKS=y+ parameter" must refer to the exclusion of the
unused backup disk. At the current location (disk layout when the backup
disk is NOT unused anymore) it does not make sense.

Move the whole part about autoexcludes just after the section where it
is mentioned that a disk was automatically excluded.