Improve the description of AUTOEXCLUDE_PATH
Separate the example of AUTOEXCLUDE_PATH handling from the treatment of
/tmp. /tmp is a bad example, because it is handled specially by the
default setting of another config variable (BACKUP_PROG_EXCLUDE).
Use /mnt as an example for AUTOEXCLUDE_PATH instead and make clear that
AUTOEXCLUDE_PATH=( /tmp ) does not exclude files under /tmp from the
backup and an independent setting is used for that purpose.