Don't look for EFI bootloader in the start of disk
EFI bootloader is not installed to the start of the disk, but to the EFI
System partition as a regular file. It is this futile to search for EFI
bootloaders in the first sectors of disks. If we find "EFI", it is only
because there is a GPT, and "EFI PART" is the GPT signature (this does
not tell anything about the presence of an EFI bootloader).

Delete the code that checked for 'Hah!IdontNeedEFI' as a special case -
it is not needed now. The code effectively skipped the check for EFI
when there was a BIOS boot partition. Its presence does not indicate
the absence of an EFI bootloader, though. We could be on an UEFI machine
with the disk partitioned in hybrid (BIOS/UEFI compatible) mode. In this
case ReaR can error out if legacy GRUB is not installed, because it does
not find any bootloader (the EFI check is skipped).

Detect EFI instead according to the USING_UEFI_BOOTLOADER variable when
no other bootloader is found.

Note that it is now perfectly legal to have USING_UEFI_BOOTLOADER=1
and BOOTLOADER neither EFI nor GRUB2-EFI nor ELILO. This will happen
with hybrid BIOS/UEFI booting: BOOTLOADER will be detected as GRUB2, but
USING_UEFI_BOOTLOADER=1.

BOOTLOADER=EFI is now intended as a fallback when no other bootloader is
found to avod an empty BOOTLOADER value.