Update framework-functions.sh
In lib/framework-functions.sh
better describe the SourceStage function implementation.
In particular tell that it breaks if a leading 3-digit number with underscore
would be used as directory name in the ReaR subdirectory structure
so e.g. a new backup method named '123_backup' is not possible
but a new backup method named '123backup' is possible.