Do not detect GRUB before including GRUB tools
When there was a file matching grub* in /boot (e.g. /boot/grub2FAIL),
the code got confused by the glob pattern intended to match /boot/grub
or /boot/grub2 and subsequently the rest of the script was skipped, as
grubdir got assigned something like "/boot/grub2 /boot/grub2FAIL", which
does not exist, so grubdir was set to /boot/grub, which does not exist
either, and grub-probe fails.

As a result, the GRUB tools were not included in the recovery image.

The code have been proceeding anyway when neither grub-probe nor
grub2-probe was found, so the tests have not been very useful.

Fix and simplify by not checking for the existence of GRUB and just
trying to include the GRUB tools always.

See the discussion in
https://github.com/rear/rear/commit/ccae513d8362078c5d4bcffe9b1167835e6449b8