Invert the meaning of exit code from is_disk_a_pv
The function is_disk_a_pv returns with 1 if disk is a PV, 0 if it is
not.

As a nonzero return code is interpreted as "false" in shell tests,
this convention is very confusing and contrary to all other is_*
functions, which return true (exit code 0) when the condition that they
check and is expressed in their name is true (cf. is_true).

Invert the exit code of is_disk_a_pv to comply with the usual convention
and adjusts comments and callers.