Warn when using BOOTLOADER = GRUB
Update comments for the code that detects if the "GRUB" value for
BOOTLOADER actually means GRUB2 and emit warnings (using LogPrintError)
about "GRUB" value to mean GRUB 2 being deprecated and the GRUB Legacy
support itself being deprecated as well.

Do not error out in this situation as erroring out near the end of
recovery only to inform about a deprecated setting would be unhelpful.