Delete comment obsolete since 67ac463446
Commit 67ac463446 ("initial attemp to guess bootloader") introduced
bootloader detection, so it has not been true since then that
"There is no guarantee that GRUB was the boot loader used originally.
One possible attempt would be to save and restore the MBR for each disk,
but this does not guarantee a correct boot order, or even a working boot
loader config" We don't save and restore the MBR for each disk, we
detect the bootloader and reinstall it, thus avoiding the latter issue.

Mention that we make "all suitable disks" bootable, not "all disks", as
we make only the / and /boot partition disks bootable and there are
further rules to skip some disks.