Consider shim when restoring EFI boot vars
UEFI_BOOTLOADER used to mean the first-stage bootloader during recovery,
either the real bootloader (e.g. GRUB), or the shim, if used - it got
set to SECURE_BOOT_BOOTLOADER in 850_save_sysfs_uefi_vars.sh.

After the cleanup of 850_save_sysfs_uefi_vars.sh UEFI_BOOTLOADER and
SECURE_BOOT_BOOTLOADER are separate and UEFI_BOOTLOADER is not used for
the shim. As a consequence, recovery now creates a boot entry pointing
to the real bootloader (GRUB) always, even if shim was used. This would
be a regression when Secure Boot is used - shim is needed to boot.

Correct this by introducing the INITIAL_UEFI_BOOTLOADER variable (not
intended to be configured by the user) that gets set either to
SECURE_BOOT_BOOTLOADER or UEFI_BOOTLOADER as appropriate and always
means the UEFI bootloader directly executed by the firmware. Use it when
restoring the EFI boot entries.

As a result, the boot entry of the recovered system now points to the shim if
SECURE_BOOT_BOOTLOADER is set and the recovered system will boot when
Secure Boot is enabled.