Fix version test in udev start by desupporting systemd < 190
/etc/scripts/system-setup.d/40-start-udev-or-load-modules.sh in the
rescue system examines systemd version in order to know whether it
should use udev or load modules manually.

Unfortunately, at least on Fedora, the systemd version can contain a
tilde (the systemd version is equal to the systemd package version,
which conforms to the Fedora conventions:
https://docs.fedoraproject.org/en-US/packaging-guidelines/Versioning/#_handling_non_sorting_versions_with_tilde_dot_and_caret )
and the bash code that compares versions chokes on it:
/etc/scripts/system-setup.d/40-start-udev-or-load-modules.sh: line 24: [[: 256~rc3: syntax error in expression (error token is "~rc3")
and then the wrong branch is taken.

This leads at least to non-working make-backup-and-restore-iso CI test
due to the ramdisk (brd) module being loaded earlier and thus not
respecting the desired parameters. (A ramdisk of a pre-determined size
is used to emulate the backup ISO in the test.) Probably has other unwanted
consequences as well.

As reasonably modern distros should have a newer systemd (even RHEL 7
has systemd 219 and Ubuntu 16.04 has 229), delete the check and error
out in a similar check during mkrescue if a too old systemd is
encountered.

The check in build/GNU/Linux/600_verify_and_adjust_udev.sh is more robust,
so it could be used as an alternative if desupporting old systemd is not
desired. The version_newer function would have to be moved to a library
available during boot, though, and the function does not really check
for a tilde in version, thus it appears to work more by accident than by
design.