Better sanity checks in SourceTrustworthy
Better sanity checks in SourceTrustworthy:
Now it only ensures the source file is a regular file (or a link to a regular file).
I.e. it skips non-regular files like directories, device nodes, or file not found.
It also returns there when no source file was specified (i.e. when $1 is empty).