Improve layout configuration part of the user guide (#3125)
* Do not refer to "restore list" in the docs

This term is never defined or used anywhere. Let's talk about the layout
file instead.

* Add "from" forgotten in the original version

* Neutral wording of the EXCLUDE_RECREATE example

Avoid the "easiest", "inadvertently" expressions.

* Don't copy EXCLUDE_BACKUP/RESTORE to layout doc

EXCLUDE_RESTORE and EXCLUDE_BACKUP are irrelevant for layout file and
disk layout recreation. They merely influence file backup or restore
(for supported backup methods). This guide is about the recreation of
the disk layout. Therefore, omit the variables from here. They are not
used in the doc anywhere anyway after showing a copy of their
description originating from default.conf. So, why to show it at all?

* layout doc: move autoexcludes description

I believe that "This behavior is controlled by the
+AUTOEXCLUDE_DISKS=y+ parameter" must refer to the exclusion of the
unused backup disk. At the current location (disk layout when the backup
disk is NOT unused anymore) it does not make sense.

Move the whole part about autoexcludes just after the section where it
is mentioned that a disk was automatically excluded.

* Do not copy almost identical layout file in docs

Show only the parts that have changed after changing something.

* Update whitespace after comment in layout examples

4d6fac780f9234b55dbb0ca7fa55b56e4fbc7157 changed the formatting of
commented entries in layout:

"Change commenting in layout file to have no space after the #."

Adapt the examples that had been generated before the change to match.

* Describe EXCLUDE_RECREATE in layout docs better

Instead of copying a snippet of default.conf, describe in prose what the
variable does and what is the syntax.

* Describe assumptions of the layout conf example

Instead of referring to what seems prudent, describe the assumptions
beyond the example explicitly. This way, the reader can understand what
we want to achieve, regardless what is their opinion on what is prudent
and what is not.

Also, describe briefly the relation between layout configuration and
backup exclusion.

* Don't claim that EXCLUDE_RECREATE is the most generic